Created a separate controller for blacklisted tokens, changed its model & moved scheduler in utils
